12.9.1-5 baseline
Former-commit-id: a1d931fd44b4a3c9fcda1a7958d7c0fbefbf930e [formerly 61f269f54c9586ae9940ef7d752bd3ec23443948] [formerly f2fac39428feb57ecf73369f09901ad397d0db06 [formerly d85b989f77196d20eb2d2a21cf4daa13d50474ae]]
Former-commit-id: f2fac39428feb57ecf73369f09901ad397d0db06
Former-commit-id: 72824561cbca2934282af23d7ce304997616cf98

+++ cave/com.raytheon.viz.ui/src/com/raytheon/viz/ui/input/PanHandler.java
@@ -211,13 +211,13 @@ public class PanHandler extends InputAdapter {
 
             if ((0 <= xMinThreshold && width >= xMaxThreshold) == false) {
                 if (((width < xMaxThreshold && theLastScreen[0] < aX) || (0 > xMinThreshold && theLastMouseX > aX)) == false) {
-                    aX2 = (int) theLastScreen[0];
+                    aX2 = theLastScreen[0];
                 }
             }
 
             if ((0 <= yMinThreshold && height >= yMaxThreshold) == false) {
                 if (((height < yMaxThreshold && theLastScreen[1] < aY) || (0 > yMinThreshold && theLastMouseY > aY)) == false) {
-                    aY2 = (int) theLastScreen[1];
+                    aY2 = theLastScreen[1];
                 }
             }
